Version 1.0
Various and plentiful improvements to frontend and backend code

+++ whatitdo/views.py
@@ -13,22 +13,6 @@ import logging
 def frontpage(request):
 	return render_to_response('frontpage.html', RequestContext(request))
 
-def register(request):
-    if request.method == 'POST':
-        form = UserCreationForm(request.POST)
-        if form.is_valid():
-            new_user = form.save()
-            return HttpResponseRedirect("/")
-    else:
-        form = UserCreationForm()
-	return render_to_response("registration/register.html", {
-        'form': form,})
-
-def login_redir(request):
-	gohomeurl = "/~" + request.user.username
-	return HttpResponseRedirect(gohomeurl)
-		
-
 def userhome(request,list_owner):
 	last_login = User.objects.get(username=list_owner).last_login
 	date_joined = User.objects.get(username=list_owner).date_joined
@@ -47,7 +31,7 @@ def userhome(request,list_owner):
 	movie_size = len(movie_list)
 	game_size = len(game_list)
 	music_size = len(music_list)
-	return render_to_response('userhome.html', {'booksize': book_size, 'tvsize': tv_size, 'moviesize': movie_size, 'gamesize': game_size, 'musicsize': music_size, 'lastlogin': last_login, 'datejoined': date_joined, 'listowner': list_owner, 'userpic':userpic, 'link1':link1, 'link2':link2, 'link3':link3}, 
+	return render_to_response('userhome.html', {'current': 'userhome', 'booksize': book_size, 'tvsize': tv_size, 'moviesize': movie_size, 'gamesize': game_size, 'musicsize': music_size, 'lastlogin': last_login, 'datejoined': date_joined, 'listowner': list_owner, 'userpic':userpic, 'link1':link1, 'link2':link2, 'link3':link3}, 
 	RequestContext(request))
 	
 	
@@ -56,7 +40,7 @@ def books(request,list_owner):
 	book_sort1 = UserProfile.objects.get(user=list_user).booksort1
 	book_sort2 = UserProfile.objects.get(user=list_user).booksort2
 	book_list = Item.objects.all().filter(item_creator=list_owner).filter(category='book').order_by(book_sort1,book_sort2,'name')
-	return render_to_response('books.html', {'booklist': book_list, 'listowner': list_owner}, 
+	return render_to_response('list.html', {'current': 'books', 'booklist': book_list, 'listowner': list_owner}, 
 	RequestContext(request))
 
 def tv(request,list_owner):
@@ -64,7 +48,7 @@ def tv(request,list_owner):
 	tv_sort1 = UserProfile.objects.get(user=list_user).tvsort1
 	tv_sort2 = UserProfile.objects.get(user=list_user).tvsort2
 	tv_list = Item.objects.all().filter(item_creator=list_owner).filter(category='tv').order_by(tv_sort1,tv_sort2,'name')
-	return render_to_response('tv.html', {'tvlist': tv_list, 'listowner': list_owner}, 
+	return render_to_response('list.html', {'current': 'tv', 'tvlist': tv_list, 'listowner': list_owner}, 
 	RequestContext(request))
 
 def movies(request,list_owner):
@@ -72,7 +56,7 @@ def movies(request,list_owner):
 	movie_sort1 = UserProfile.objects.get(user=list_user).moviesort1
 	movie_sort2 = UserProfile.objects.get(user=list_user).moviesort2
 	movie_list = Item.objects.all().filter(item_creator=list_owner).filter(category='movie').order_by(movie_sort1,movie_sort2,'name')
-	return render_to_response('movies.html', {'movielist': movie_list, 'listowner': list_owner}, 
+	return render_to_response('list.html', {'current': 'movies', 'movielist': movie_list, 'listowner': list_owner}, 
 	RequestContext(request))
 
 def games(request,list_owner):
@@ -80,7 +64,7 @@ def games(request,list_owner):
 	game_sort1 = UserProfile.objects.get(user=list_user).gamesort1
 	game_sort2 = UserProfile.objects.get(user=list_user).gamesort2
 	game_list = Item.objects.all().filter(item_creator=list_owner).filter(category='game').order_by(game_sort1,game_sort2,'name')
-	return render_to_response('games.html', {'gamelist': game_list, 'listowner': list_owner}, 
+	return render_to_response('list.html', {'current': 'games', 'gamelist': game_list, 'listowner': list_owner}, 
 	RequestContext(request))
 
 def music(request,list_owner):
@@ -88,7 +72,7 @@ def music(request,list_owner):
 	music_sort1 = UserProfile.objects.get(user=list_user).musicsort1
 	music_sort2 = UserProfile.objects.get(user=list_user).musicsort2
 	music_list = Item.objects.all().filter(item_creator=list_owner).filter(category='music').order_by(music_sort1,music_sort2,'name')
-	return render_to_response('music.html', {'musiclist': music_list, 'listowner': list_owner}, 
+	return render_to_response('list.html', {'current': 'music', 'musiclist': music_list, 'listowner': list_owner}, 
 	RequestContext(request))
 
 @login_required
@@ -106,7 +90,7 @@ def editbook(request,list_owner,pk):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem(instance=item_to_edit)
-	return render_to_response('editbook.html', {'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'books', 'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
 
 @login_required
 def edittv(request,list_owner,pk):
@@ -123,7 +107,7 @@ def edittv(request,list_owner,pk):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem(instance=item_to_edit)
-	return render_to_response('edittv.html', {'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'tv', 'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
 
 @login_required
 def editmovie(request,list_owner,pk):
@@ -140,7 +124,7 @@ def editmovie(request,list_owner,pk):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem(instance=item_to_edit)
-	return render_to_response('editmovie.html', {'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'movies', 'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
 
 @login_required
 def editgame(request,list_owner,pk):
@@ -157,7 +141,7 @@ def editgame(request,list_owner,pk):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem(instance=item_to_edit)
-	return render_to_response('editgame.html', {'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'games', 'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
 
 @login_required
 def editmusic(request,list_owner,pk):
@@ -174,7 +158,7 @@ def editmusic(request,list_owner,pk):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem(instance=item_to_edit)
-	return render_to_response('editmusic.html', {'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'music', 'listowner': list_owner, 'item_to_edit': item_to_edit, 'form':form}, RequestContext(request))
 
 @login_required
 def newbook(request,list_owner):
@@ -190,7 +174,7 @@ def newbook(request,list_owner):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem()
-	return render_to_response('editbook.html', {'listowner': list_owner, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'books', 'listowner': list_owner, 'form':form}, RequestContext(request))
 
 @login_required
 def newtv(request,list_owner):
@@ -206,7 +190,7 @@ def newtv(request,list_owner):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem()
-	return render_to_response('edittv.html', {'listowner': list_owner, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'tv', 'listowner': list_owner, 'form':form}, RequestContext(request))
 
 @login_required
 def newmovie(request,list_owner):
@@ -222,7 +206,7 @@ def newmovie(request,list_owner):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem()
-	return render_to_response('editmovie.html', {'listowner': list_owner, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'movies', 'listowner': list_owner, 'form':form}, RequestContext(request))
 
 @login_required
 def newgame(request,list_owner):
@@ -238,7 +222,7 @@ def newgame(request,list_owner):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem()
-	return render_to_response('editgame.html', {'listowner': list_owner, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'games', 'listowner': list_owner, 'form':form}, RequestContext(request))
 
 @login_required
 def newmusic(request,list_owner):
@@ -254,10 +238,13 @@ def newmusic(request,list_owner):
 			return HttpResponseRedirect(returnurl)
 	else:
 		form = EditItem()
-	return render_to_response('editmusic.html', {'listowner': list_owner, 'form':form}, RequestContext(request))
+	return render_to_response('edit.html', {'current': 'music', 'listowner': list_owner, 'form':form}, RequestContext(request))
 
 @login_required
 def delbook(request,list_owner,pk):
+	if request.user.username != list_owner:
+		gohomeurl = "/~" + request.user.username
+		return HttpResponseRedirect(gohomeurl)
 	list_user = User.objects.get(username=list_owner)
 	item_to_edit = Item.objects.get(pk=pk)
 	item_to_edit.delete()
@@ -266,6 +253,9 @@ def delbook(request,list_owner,pk):
 
 @login_required
 def delgame(request,list_owner,pk):
+	if request.user.username != list_owner:
+		gohomeurl = "/~" + request.user.username
+		return HttpResponseRedirect(gohomeurl)
 	list_user = User.objects.get(username=list_owner)
 	item_to_edit = Item.objects.get(pk=pk)
 	item_to_edit.delete()
@@ -274,6 +264,9 @@ def delgame(request,list_owner,pk):
 
 @login_required
 def delmovie(request,list_owner,pk):
+	if request.user.username != list_owner:
+		gohomeurl = "/~" + request.user.username
+		return HttpResponseRedirect(gohomeurl)
 	list_user = User.objects.get(username=list_owner)
 	item_to_edit = Item.objects.get(pk=pk)
 	item_to_edit.delete()
@@ -282,6 +275,9 @@ def delmovie(request,list_owner,pk):
 
 @login_required
 def deltv(request,list_owner,pk):
+	if request.user.username != list_owner:
+		gohomeurl = "/~" + request.user.username
+		return HttpResponseRedirect(gohomeurl)
 	list_user = User.objects.get(username=list_owner)
 	item_to_edit = Item.objects.get(pk=pk)
 	item_to_edit.delete()
@@ -290,22 +286,33 @@ def deltv(request,list_owner,pk):
 
 @login_required
 def delmusic(request,list_owner,pk):
+	if request.user.username != list_owner:
+		gohomeurl = "/~" + request.user.username
+		return HttpResponseRedirect(gohomeurl)
 	list_user = User.objects.get(username=list_owner)
 	item_to_edit = Item.objects.get(pk=pk)
 	item_to_edit.delete()
 	returnurl = "/~" + list_owner + "/music"
 	return HttpResponseRedirect(returnurl)
 
+
+
 @login_required
 def userpic(request,list_owner):
+	if request.user.username != list_owner:
+		redirect = "/~" + request.user.username + "/userpic"
+		return HttpResponseRedirect(redirect)
 	current_user = User.objects.get(pk=request.user.id)
 	current_profile = current_user.get_profile()
 	current_userpic = UserProfile.objects.get(user_id=current_user).userpic
 	if request.method == 'POST':
 		form = UserpicUpload(request.POST, request.FILES, instance=current_profile)
 		if form.is_valid():
-			current_profile.userpic=request.FILES['userpic']
-			current_profile.save()
+			try:
+				current_profile.userpic=request.FILES['userpic']
+				current_profile.save()
+			except:
+				pass
 			gohomeurl = "/~" + request.user.username
 			return HttpResponseRedirect(gohomeurl)
 	else:
@@ -314,6 +321,9 @@ def userpic(request,list_owner):
 
 @login_required
 def editlinks(request,list_owner):
+	if request.user.username != list_owner:
+		redirect = "/~" + request.user.username + "/editlinks"
+		return HttpResponseRedirect(redirect)
 	current_user = User.objects.get(pk=request.user.id)
 	current_profile = current_user.get_profile()
 	current_link1 = UserProfile.objects.get(user_id=current_user).link1
@@ -332,8 +342,12 @@ def editlinks(request,list_owner):
 		form = EditLinks()
 	return render_to_response('editlinks.html', {'listowner': list_owner, 'form':form, 'currentlink1':current_link1, 'currentlink2':current_link2, 'currentlink3':current_link3}, RequestContext(request))
 
+
 @login_required
 def listsort(request,list_owner):
+	if request.user.username != list_owner:
+		redirect = "/~" + request.user.username + "/listsort"
+		return HttpResponseRedirect(redirect)
 	current_user = User.objects.get(pk=request.user.id)
 	current_profile = current_user.get_profile()
 	tvsort1 = UserProfile.objects.get(user_id=current_user).tvsort1
@@ -380,3 +394,18 @@ def contact(request):
 
 def contact_thanks(request): 
 	return render_to_response('contact_thanks.html', RequestContext(request))
+
+
+def register(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            new_user = form.save()
+            return HttpResponseRedirect("/")
+    else:
+        form = UserCreationForm()
+	return render_to_response("registration/register.html", {'form': form})
+
+def login_redir(request):
+	gohomeurl = "/~" + request.user.username
+	return HttpResponseRedirect(gohomeurl)
